# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BookStore is a lightweight bookstore web API built with Golang (Gin-Gonic).",
        "title": "Book API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/order": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all Orders for User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessDTO"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Place an Order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/order/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Order by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/order/{id}/payment": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Order Payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment Info",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Payment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/order/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Order Status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status Info",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Status"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.ErrorDTO": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                }
            }
        },
...
        "main.Payment": {
            "type": "object",
            "required": [
                "method"
            ],
            "properties": {
                "method": {
                    "type": "string",
                    "enum": [
                        "UPI",
                        "Card",
                        "COD"
                    ]
                },
                "paid": {
                    "type": "boolean"
                },
                "paid_on": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                }
            }
        },
...
        "main.Status": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "reason": {
                    "type": "string"
                },
                "shipping_carrier": {
                    "type": "string"
                },
                "tracking_number": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "shipped",
                        "delivered"
                    ]
                }
            }
        },
        "main.SuccessDTO": {
            "type": "object",
            "properties": {
                "response": {},
                "status_code": {
                    "type": "string"
                },
                "status_message": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}